requirements.txt
tweepy
art
py3cw
flashtext
urllib3
python-telegram-bot
uptime

